fix a couple of Clang 10 warnings

+++ source/distributed/cell_weights.cc
@@ -86,7 +86,9 @@ namespace parallel
         std::trunc(coefficients.first *
                    std::pow(future_fe.dofs_per_cell, coefficients.second));
 
-      Assert(result >= 0 && result <= std::numeric_limits<unsigned int>::max(),
+      Assert(result >= 0. &&
+               result <=
+                 static_cast<float>(std::numeric_limits<unsigned int>::max()),
              ExcMessage(
                "Cannot cast determined weight for this cell to unsigned int!"));
 
@@ -109,7 +111,9 @@ namespace parallel
         result += pair.first * std::pow(future_fe.dofs_per_cell, pair.second);
       result = std::trunc(result);
 
-      Assert(result >= 0 && result <= std::numeric_limits<unsigned int>::max(),
+      Assert(result >= 0. &&
+               result <=
+                 static_cast<float>(std::numeric_limits<unsigned int>::max()),
              ExcMessage(
                "Cannot cast determined weight for this cell to unsigned int!"));
